backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
langchain
langchain-openai
langchain-core
python-dotenv
httpx
python-jose[cryptography]
bcrypt
python-multipart
openpyxl
reportlab

